+++ controllerhawk.js
@@ -35,9 +35,188 @@ function _defineProperty(obj, key, value) {
   return obj;
 }
 
+var Keybind =
+/*#__PURE__*/
+function () {
+  /**
+   * The buttons that are assigned to this keybind.
+   * 
+   * @private
+   * 
+   * @property {KeybindObject}
+   */
+
+  /**
+   * The callback method to run when this keybind is used.
+   * 
+   * @private
+   * 
+   * @property {Function}
+   * 
+   * @default this.noop
+   */
+
+  /**
+   * The timestamp of the last time this keybind was used.
+   * 
+   * @private
+   * 
+   * @property {number}
+   * 
+   * @default 0
+   */
+
+  /**
+   * @param {KeybindObject} buttons The buttons to bind to this keybind.
+   */
+  function Keybind(buttons) {
+    _classCallCheck(this, Keybind);
+
+    _defineProperty(this, "_buttons", void 0);
+
+    _defineProperty(this, "_action", this._noop);
+
+    _defineProperty(this, "_lastUsed", 0);
+
+    this._buttons = buttons;
+  }
+  /**
+   * Returns the buttons for this keybind.
+   * 
+   * @returns {Array<number>}
+   */
+
+
+  _createClass(Keybind, [{
+    key: "action",
+
+    /**
+     * Sets the function to run when this keybind is used.
+     * 
+     * @param {Function} fn The function to run when this keybind is used.
+     * 
+     * @returns {Keybind} Returns this for chaining.
+     */
+    value: function action(fn) {
+      this._action = fn;
+      return this;
+    }
+    /**
+     * Runs the function associated with this keybind.
+     * 
+     * @param {number} time The timestamp passed from the game clock.
+     */
+
+  }, {
+    key: "run",
+    value: function run(time) {
+      this._action();
+
+      this._lastUsed = time;
+    }
+    /**
+     * An empty function to use as a default value for the action of this keybind.
+     * 
+     * @private
+     */
+
+  }, {
+    key: "_noop",
+    value: function _noop() {}
+  }, {
+    key: "buttons",
+    get: function get() {
+      return this._buttons;
+    }
+    /**
+     * Returns the last time that this keybind was used.
+     * 
+     * @returns {number}
+     */
+
+  }, {
+    key: "lastUsed",
+    get: function get() {
+      return this._lastUsed;
+    }
+  }]);
+
+  return Keybind;
+}();
+
+/**
+ * Maps the standard gamepad to a PS4 controller.
+ * 
+ * https://www.w3.org/TR/gamepad/#remapping
+ */
+
+var ps4 = {
+  X: 0,
+  CROSS: 0,
+  CIRCLE: 1,
+  SQUARE: 2,
+  TRIANGLE: 3,
+  LEFT_BUMPER: 4,
+  RIGHT_BUMPER: 5,
+  LEFT_TRIGGER: 6,
+  RIGHT_TRIGGER: 7,
+  SELECT: 8,
+  START: 9,
+  LEFT_ANALOG: 10,
+  RIGHT_ANALOG: 11,
+  DPAD_UP: 12,
+  DPAD_BOTTOM: 13,
+  DPAD_LEFT: 14,
+  DPAD_RIGHT: 15,
+  MIDDLE: 16
+};
+
+/**
+ * Maps the standard gamepad to a Xbox One controller.
+ * 
+ * https://www.w3.org/TR/gamepad/#remapping
+ */
+
+var xbox = {
+  A: 0,
+  B: 1,
+  X: 2,
+  Y: 3,
+  LEFT_BUMPER: 4,
+  RIGHT_BUMPER: 5,
+  LEFT_TRIGGER: 6,
+  RIGHT_TRIGGER: 7,
+  SELECT: 8,
+  START: 9,
+  LEFT_ANALOG: 10,
+  RIGHT_ANALOG: 11,
+  DPAD_UP: 12,
+  DPAD_BOTTOM: 13,
+  DPAD_LEFT: 14,
+  DPAD_RIGHT: 15,
+  MIDDLE: 16
+};
+
+/**
+ * Imports all of the individual mappings and provides a way to access any supported controller layout.
+ */
+
+var controllers = {
+  /**
+   * The layout for the ps4 controller.
+   */
+  PS4: ps4,
+
+  /**
+   * The layout for the general xbox/xbox360 controller.
+   */
+  XBOX: xbox
+};
+
 /**
  * Makes it easy for you to map controllers for your game.
  */
+
 var ControllerHawk =
 /*#__PURE__*/
 function () {
@@ -48,25 +227,83 @@ function () {
    * 
    * @property {Controllers}
    */
+
+  /**
+   * A reference to the controllers and their buttons that can be used in keybinds.
+   * 
+   * @private
+   * 
+   * @property {controllers}
+   */
+
+  /**
+   * A reference to all of the keybinds that have been created.
+   * 
+   * @private
+   * 
+   * @property {Array<Keybind>}
+   */
+
+  /**
+   * A reference to the buttons on the controller currently being pressed.
+   * 
+   * @private
+   * 
+   * @property {*}
+   */
   function ControllerHawk() {
     _classCallCheck(this, ControllerHawk);
 
     _defineProperty(this, "_controllers", {});
 
+    _defineProperty(this, "_BUTTONS", controllers);
+
+    _defineProperty(this, "_keybinds", []);
+
+    _defineProperty(this, "_pressed", {});
+
     this._boot();
   }
   /**
-   * Maps a contoller button bind to an action.
-   */
-
-  /**
-   * Adds the necessary event listeners detecting a gamepad connecting and disconnecting.
+   * Returns the controller mappings.
    * 
-   * @private
+   * @returns {controller}
    */
 
 
   _createClass(ControllerHawk, [{
+    key: "keybind",
+
+    /**
+     * Creates a new keybind with the specified buttons on a supported controller layout.
+     * 
+     * @param {...string} buttons One or more buttons from the `CONTROLLER` property to attach to this keybind.
+     * 
+     * @returns {Keybind} Returns the newly created keybind.
+     */
+    value: function keybind() {
+      for (var _len = arguments.length, buttons = new Array(_len), _key = 0; _key < _len; _key++) {
+        buttons[_key] = arguments[_key];
+      }
+
+      if (!buttons) {
+        console.warn('At least one button must be provided to create a keybind.');
+        return;
+      }
+
+      var keybind = new Keybind(buttons);
+
+      this._keybinds.push(keybind);
+
+      return keybind;
+    }
+    /**
+     * Adds the necessary event listeners detecting a gamepad connecting and disconnecting.
+     * 
+     * @private
+     */
+
+  }, {
     key: "_boot",
     value: function _boot() {
       var _this = this;
@@ -90,12 +327,9 @@ function () {
   }, {
     key: "_onconnect",
     value: function _onconnect(gamepad) {
-      var _this2 = this;
-
       this._controllers[gamepad.index] = gamepad;
-      requestAnimationFrame(function () {
-        return _this2._update();
-      });
+
+      this._update(0);
     }
     /**
      * Removes the disconnected controller from the `_controllers` property.
@@ -113,19 +347,69 @@ function () {
     /**
      * Checks to see if there are any gamepads that need to be added and performs controller actions.
      * 
+     * @param {number} time The time passed from the game clock.
+     * 
      * @private
      */
 
   }, {
     key: "_update",
-    value: function _update() {
-      var _this3 = this;
+    value: function _update(time) {
+      var _this2 = this;
 
       this._scan();
 
-      requestAnimationFrame(function () {
-        return _this3._update();
+      this._updatePressed();
+
+      this._keybinds.map(function (keybind) {
+        var _iteratorNormalCompletion = true;
+        var _didIteratorError = false;
+        var _iteratorError = undefined;
+
+        try {
+          for (var _iterator = keybind.buttons[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {
+            var button = _step.value;
+            if (!_this2._pressed[button]) return;
+          }
+        } catch (err) {
+          _didIteratorError = true;
+          _iteratorError = err;
+        } finally {
+          try {
+            if (!_iteratorNormalCompletion && _iterator["return"] != null) {
+              _iterator["return"]();
+            }
+          } finally {
+            if (_didIteratorError) {
+              throw _iteratorError;
+            }
+          }
+        }
+
+        keybind.run(time);
       });
+
+      requestAnimationFrame(function (time) {
+        return _this2._update(time);
+      });
+    }
+    /**
+     * Updates the status of what buttons are pressed.
+     * 
+     * @private
+     */
+
+  }, {
+    key: "_updatePressed",
+    value: function _updatePressed() {
+      for (var c in this._controllers) {
+        var controller = this._controllers[c];
+
+        for (var i = 0; i < controller.buttons.length; ++i) {
+          var button = controller.buttons[i];
+          if (button.pressed) this._pressed[i] = true;else this._pressed[i] = false;
+        }
+      }
     }
     /**
      * Scans for new controllers and if so it adds them to the `_controllers` property or updates the current references.
@@ -136,13 +420,18 @@ function () {
   }, {
     key: "_scan",
     value: function _scan() {
-      var _this4 = this;
+      var _this3 = this;
 
       var gamepads = navigator.getGamepads();
       gamepads.map(function (gamepad) {
         if (!gamepad) return;
-        if (!(gamepad.index in _this4._controllers)) _this4._onconnect(gamepad);else _this4._controllers[gamepad.index] = gamepad;
+        if (!(gamepad.index in _this3._controllers)) _this3._onconnect(gamepad);else _this3._controllers[gamepad.index] = gamepad;
       });
+    }
+  }, {
+    key: "BUTTONS",
+    get: function get() {
+      return this._BUTTONS;
     }
   }]);