+++ src/index.ts
@@ -1,6 +1,7 @@
 'use strict'
 
-import Controllers from './interfaces/Controllers';
+import Keybind from './keybind/Keybind';
+import controllers from './mapping/controllers';
 
 /**
  * Makes it easy for you to map controllers for your game.
@@ -13,15 +14,65 @@ export default class ControllerHawk {
    * 
    * @property {Controllers}
    */
-  private _controllers: Controllers = {};
+  private _controllers: any = {};
+
+  /**
+   * A reference to the controllers and their buttons that can be used in keybinds.
+   * 
+   * @private
+   * 
+   * @property {controllers}
+   */
+  private _BUTTONS: any = controllers;
+
+  /**
+   * A reference to all of the keybinds that have been created.
+   * 
+   * @private
+   * 
+   * @property {Array<Keybind>}
+   */
+  private _keybinds: Array<Keybind> = [];
+
+  /**
+   * A reference to the buttons on the controller currently being pressed.
+   * 
+   * @private
+   * 
+   * @property {*}
+   */
+  private _pressed: any = {};
 
   constructor() {
     this._boot();
   }
 
   /**
-   * Maps a contoller button bind to an action.
+   * Returns the controller mappings.
+   * 
+   * @returns {controller}
    */
+  get BUTTONS() { return this._BUTTONS; }
+
+  /**
+   * Creates a new keybind with the specified buttons on a supported controller layout.
+   * 
+   * @param {...string} buttons One or more buttons from the `CONTROLLER` property to attach to this keybind.
+   * 
+   * @returns {Keybind} Returns the newly created keybind.
+   */
+  keybind(...buttons: Array<number>): (Keybind | undefined) {
+    if (!buttons) {
+      console.warn('At least one button must be provided to create a keybind.');
+      return;
+    }
+
+    const keybind: Keybind = new Keybind(buttons);
+
+    this._keybinds.push(keybind);
+
+    return keybind;
+  }
 
   /**
    * Adds the necessary event listeners detecting a gamepad connecting and disconnecting.
@@ -44,7 +95,7 @@ export default class ControllerHawk {
   private _onconnect(gamepad: Gamepad) {
     this._controllers[gamepad.index] = gamepad;
 
-    requestAnimationFrame(() => this._update());
+    this._update(0);
   }
 
   /**
@@ -61,12 +112,43 @@ export default class ControllerHawk {
   /**
    * Checks to see if there are any gamepads that need to be added and performs controller actions.
    * 
+   * @param {number} time The time passed from the game clock.
+   * 
    * @private
    */
-  private _update() {
+  private _update(time: number) {
     this._scan();
 
-    requestAnimationFrame(() => this._update());
+    this._updatePressed();
+
+    this._keybinds.map((keybind: Keybind) => {
+
+      for (let button of keybind.buttons) {
+        if (!this._pressed[button]) return;
+      }
+
+      keybind.run(time);
+    });
+
+    requestAnimationFrame((time: number) => this._update(time));
+  }
+
+  /**
+   * Updates the status of what buttons are pressed.
+   * 
+   * @private
+   */
+  private _updatePressed() {
+    for (const c in this._controllers) {
+      const controller: any = this._controllers[c];
+
+      for (let i: number = 0; i < controller.buttons.length; ++i) {
+        const button: GamepadButton = controller.buttons[i];
+
+        if (button.pressed) this._pressed[i] = true;
+        else this._pressed[i] = false;
+      }
+    }
   }
 
   /**
@@ -75,7 +157,7 @@ export default class ControllerHawk {
    * @private
    */
   private _scan() {
-    const gamepads: Array<Gamepad|null> = navigator.getGamepads();
+    const gamepads: Array<Gamepad | null> = navigator.getGamepads();
 
     gamepads.map(gamepad => {
       if (!gamepad) return;