+++ src/interfaces/Controllers.ts
@@ -1,8 +1,0 @@
-'use strict'
-
-/**
- * Defines the structure of the object that holds connected controllers.
- */
-export default interface Controllers {
-  [key: number]: Gamepad;
-}