+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.0.2 / 2020-02-20
+- [FEATURE] Added ability to add keybinds.
+- [FEATURE] Added controller mappings for ps4, xbox, and generic controllers.
+- [DOCS] Started documenting keybinds.
+
 ## 0.0.1 / 2020-02-14
 - Initial commit